+++ app/controllers/shopping_lists_controller.rb
@@ -1,0 +1,3 @@
+class ShoppingListsController < ApplicationController
+  include IngredientsAggregator
+end
