+++ app/helpers/shopping_lists_helper.rb
@@ -1,0 +1,2 @@
+module ShoppingListsHelper
+end
