+++ test/controllers/shopping_lists_controller_test.rb
@@ -1,0 +1,7 @@
+require "test_helper"
+
+class ShoppingListsControllerTest < ActionDispatch::IntegrationTest
+  # test "the truth" do
+  #   assert true
+  # end
+end
